feat: add ConnectTimeout default and config override
Specifies the timeout (in seconds) used when connecting to the SSH
server, instead of using the default system TCP timeout. This timeout is
applied both to establishing the connection and to performing the
initial SSH protocol handshake and key exchange.

This addition of this option prevents any unreasonable hang up issues
where the ssh connection cannot be established.